Microservice Documentation/Helm Deploy/NodePort

+++ src/tasks/utils/sample_utils.cr
@@ -19,6 +19,32 @@ def cnf_conformance_yml
   Totem.from_file "./#{cnf_conformance}"
 end
 
+def get_parsed_cnf_conformance_yml(args)
+  if args.named.keys.includes? "yml-file"
+    yml_file = args.named["yml-file"].as(String)
+    return Totem.from_file "#{yml_file}"
+  else
+    cnf_conformance = `find cnfs/* -name "cnf-conformance.yml"`.split("\n")[0]
+    if cnf_conformance.empty?
+      raise "No cnf_conformance.yml found! Did you run the setup task?"
+    end
+    Totem.from_file "./#{cnf_conformance}"
+  end
+end
+
+def cnf_conformance_yml_file_path(args)
+  if args.named.keys.includes? "yml-file"
+    yml_file = args.named["yml-file"].as(String)
+    cnf_conformance = File.expand_path(yml_file).split("/")[0..-2].reduce(""){|x,acc| x == "" ? "/" : x + acc + "/"}
+  else
+    cnf_conformance = `find cnfs/* -name "cnf-conformance.yml"`.split("\n")[0]
+    if cnf_conformance.empty?
+      raise "No cnf_conformance.yml found! Did you run the setup task?"
+    end
+  end
+  cnf_conformance
+end 
+
 def final_cnf_results_yml
   results_file = `find ./* -name "cnf-conformance-results-*.yml"`.split("\n")[-2].gsub("./", "")
   if results_file.empty?
@@ -53,6 +79,20 @@ def cnf_conformance_dir(source_dir)
     raise "No directory named #{source_dir} found! Did you run the setup task?"
   end
   cnf_conformance.split("/")[-1] 
+end
+
+def get_cnf_conformance_dir(args)
+  if args.named.keys.includes? "yml-file"
+    yml_file = args.named["yml-file"].as(String)
+    config = Totem.from_file "#{yml_file}"
+    config.get("helm_directory").as_s
+  else
+    cnf_conformance = `find cnfs/* -name "cnf-conformance.yml"`.split("\n")[0]
+    if cnf_conformance.empty?
+      raise "No cnf_conformance.yml found! Did you run the setup task?"
+    end
+    cnf_conformance.split("/")[-2]
+  end
 end
 
 def sample_conformance_yml(sample_dir)
@@ -127,11 +167,11 @@ def sample_destination_dir(sample_source_dir)
   "#{current_dir}/#{CNF_DIR}/#{short_sample_dir(sample_source_dir)}"
 end
 
-def helm_repo_add(helm_repo_name=nil, helm_repo_url=nil)
+def helm_repo_add(helm_repo_name=nil, helm_repo_url=nil, args : Sam::Args=Sam::Args.new)
   ret = false
   if helm_repo_name == nil || helm_repo_url == nil
-    config = cnf_conformance_yml
-    current_dir = FileUtils.pwd 
+    config = get_parsed_cnf_conformance_yml(args)
+    current_dir = FileUtils.pwd
     helm = "#{current_dir}/#{TOOLS_DIR}/helm/linux-amd64/helm"
     helm_repo_name = config.get("helm_repository.name").as_s?
     helm_repo_url = config.get("helm_repository.repo_url").as_s?
@@ -143,7 +183,7 @@ def helm_repo_add(helm_repo_name=nil, helm_repo_url=nil)
     else
       ret = false
     end
-  else 
+  else
     ret = false
   end
   ret
@@ -247,7 +287,7 @@ def tools_helm
   helm = "#{current_dir}/#{TOOLS_DIR}/helm/linux-amd64/helm"
 end
 
-def sample_cleanup(sample_dir, verbose=true)
+def sample_cleanup(sample_dir, force=false, verbose=true)
   config = sample_conformance_yml(sample_dir)
   release_name = config.get("release_name").as_s 
 
@@ -255,11 +295,15 @@ def sample_cleanup(sample_dir, verbose=true)
   helm = "#{current_dir}/#{TOOLS_DIR}/helm/linux-amd64/helm"
   puts helm if verbose 
   destination_cnf_dir = "#{current_dir}/#{CNF_DIR}/#{short_sample_dir(sample_dir)}"
-  rm = `rm -rf #{destination_cnf_dir}`
-  puts rm if verbose
-  helm_uninstall = `#{helm} uninstall #{release_name}`
-  ret = $?
-  puts helm_uninstall if verbose
+  dir_exists = File.directory?(destination_cnf_dir)
+  ret = true
+  if dir_exists || force == true
+    rm = `rm -rf #{destination_cnf_dir}`
+    puts rm if verbose
+    helm_uninstall = `#{helm} uninstall #{release_name}`
+    ret = $?.success?
+    puts helm_uninstall if verbose
+  end
   ret
 end
